will add scenario that verifies mysql resiliency on kubernetes volume
The folder structure is arranged as per go standard.
The cmd folder is expected to consist of test cases.
Each test case will have its own folder.

Each test case dir is expected to consist of files such as:
- e2e.feature
- e2e_test.go
- README.md
- Dockerfile
- etc

Overall idea is to have a test case explained as a bunch of
scenarios. A scenario is an example of the system's behavior
from one or more users' perspectives.

In this commit, the scenario has been implemented using
https://github.com/DATA-DOG/godog/ library.

Signed-off-by: AmitKumarDas <amit.das@openebs.io>